Do not pull out ComboBox example data
