+++ src/combobox/example/index.ts
@@ -7,7 +7,59 @@ import ComboBox from '../ComboBox';
 import ResultItem from '../ResultItem';
 import ResultMenu from '../ResultMenu';
 import dojoTheme from '../../themes/dojo/theme';
-import data from './data';
+
+const data = [
+	{ value: 'Maine' },
+	{ value: 'New Hampshire' },
+	{ value: 'Vermont' },
+	{ value: 'Massachusetts' },
+	{ value: 'Connecticut' },
+	{ value: 'Rhode Island' },
+	{ value: 'New York' },
+	{ value: 'New Jersey' },
+	{ value: 'Pennsylvania' },
+	{ value: 'Delaware' },
+	{ value: 'Maryland' },
+	{ value: 'Virginia' },
+	{ value: 'Florida' },
+	{ value: 'Texas' },
+	{ value: 'Kentucky' },
+	{ value: 'Tennessee' },
+	{ value: 'North Carolina' },
+	{ value: 'South Carolina' },
+	{ value: 'Georgia' },
+	{ value: 'Alabama' },
+	{ value: 'Mississippi' },
+	{ value: 'Arkansas' },
+	{ value: 'Louisiana' },
+	{ value: 'Missouri' },
+	{ value: 'Oklahoma' },
+	{ value: 'Ohio' },
+	{ value: 'Nebraska' },
+	{ value: 'Michigan' },
+	{ value: 'Indiana' },
+	{ value: 'Wisconsin' },
+	{ value: 'Illinois' },
+	{ value: 'Minnesota' },
+	{ value: 'Iowa' },
+	{ value: 'North Dakota' },
+	{ value: 'South Dakota' },
+	{ value: 'Kansas' },
+	{ value: 'Colorado' },
+	{ value: 'New Mexico' },
+	{ value: 'Arizona' },
+	{ value: 'Nevada' },
+	{ value: 'California' },
+	{ value: 'Wyoming' },
+	{ value: 'Montana' },
+	{ value: 'Utah' },
+	{ value: 'Idaho' },
+	{ value: 'Washington' },
+	{ value: 'Oregon' },
+	{ value: 'Alaska' },
+	{ value: 'Hawaii' },
+	{ value: 'West Virginia' }
+];
 
 class CustomResultItem extends ResultItem {
 	renderResult(result: any) {
